+++ libft/includes/libft.h
@@ -6,7 +6,7 @@
 /*   By: nide-mel <nide-mel@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/13 14:09:34 by nide-mel          #+#    #+#             */
-/*   Updated: 2021/09/19 20:56:50 by nide-mel         ###   ########.fr       */
+/*   Updated: 2021/09/19 23:46:25 by nide-mel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,7 +30,7 @@ void		*ft_memccpy(void *dest, const void *str, int c, size_t n);
 void		*ft_memmove(void *dest, const void *str, size_t n);
 void		*ft_memchr(const void *str, int c, size_t n);
 int			ft_memcmp(const void *str1, const void *str2, size_t n);
-size_t		ft_strlen(const char *str);
+int			ft_strlen(const char *str);
 size_t		ft_strlcpy(char *dest, const char *src, size_t n);
 size_t		ft_strlcat(char *dest, const char *src, size_t n);
 int			ft_strchr(char *s, char c);
